requirements.txt
beautifulsoup4==4.13.4
fastapi==0.115.12
Flask==3.1.0
langchain>=0.4.0
langchain_community==0.3.21
langchain_core==0.3.68
langchain_openai==0.3.14
langchain_tools==0.1.34
langgraph==0.5.3
langgraph-prebuilt==0.5.2
langgraph-supervisor==0.0.27
langserve==0.3.1
Pillow==11.2.1
pydantic==2.11.3
python-dotenv==1.1.0
Requests==2.32.3
typing_extensions==4.13.2
uvicorn==0.34.1
pytz==2024.1
langsmith==0.4.5
alibabacloud_green20220302==2.21.2
alibabacloud_tea_openapi==0.4.3
ragas==0.3.0
gradio==4.32.0
faiss-cpu==1.8.0.post1
dashscope==1.20.14
openai==1.55.3
httpx==0.27.0
llama-index-vector-stores-faiss==0.1.2
llama-index-embeddings-dashscope==0.1.4
llama-index-readers-file==0.1.33
docx2txt==0.8
openpyxl==3.1.5
llama-index-core==0.10.67
llama-index-llms-dashscope==0.1.2
llama-index-readers-dashscope==0.1.2
llama-index-llms-openai-like==0.1.3
llama-index-postprocessor-dashscope-rerank-custom==0.1.0
simplejson==3.19.3
matplotlib==3.9.2
alibabacloud_bailian20231229==1.8.2
pandas==2.0.3
oss2==2.19.0
lagent==0.1.2
mmengine==0.10.3
ipywidgets==8.1.7
modelscope
